+++ src/main/java/com/web_project/shop/service/StudentService.java
@@ -8,11 +8,11 @@ public interface StudentService {
 
     public List<StudentModel> findAllStudents();
 
-    public StudentModel findStudentById(int id);
+    public StudentModel findStudentById(Long id);
 
     public StudentModel addStudent(StudentModel student);
 
     public StudentModel updateStudent(StudentModel student);
 
-    public void deleteStudent(int id);
+    public void deleteStudent(Long id);
 }
